requirements.txt
python-telegram-bot==21.6
requests==2.32.3
pytz==2024.2
